refactoring of server cancellation request
